# docs/swagger.yaml
basePath: /api
definitions:
  models.AckDTO:
    properties:
      claimToken:
        type: string
      errors:
        items:
          $ref: '#/definitions/models.ItemError'
        type: array
      processedCount:
        minimum: 0
        type: integer
    required:
    - claimToken
    type: object
...
  models.AppError:
    properties:
      codigo:
        type: integer
      mensagemErro:
        type: string
    type: object
  models.BatchDTO:
    properties:
      batchNumber:
        minimum: 1
        type: integer
      id:
        type: string
      isLast:
        type: boolean
      items:
        items:
          type: object
        maxItems: 1000
        type: array
      totalItems:
        minimum: 0
        type: integer
    required:
    - id
    type: object
...
  models.FailDTO:
    properties:
      claimToken:
        type: string
      error:
        maxLength: 2000
        type: string
    required:
    - claimToken
    - error
    type: object
...
  models.HeartbeatDTO:
    properties:
      customers:
        minimum: 0
        type: integer
      databaseOk:
        type: boolean
      employees:
        minimum: 0
        type: integer
      errorCount:
        minimum: 0
        type: integer
      pendingCount:
        minimum: 0
        type: integer
      products:
        minimum: 0
        type: integer
      status:
        enum:
        - ONLINE
        - OFFLINE
        - DEGRADED
        type: string
      suppliers:
        minimum: 0
        type: integer
      version:
        maxLength: 50
        type: string
    required:
    - status
    - version
    type: object
...
  models.ItemError:
    properties:
      error:
        type: string
      errorCode:
        type: string
      itemKey:
        type: string
    type: object
...
  models.Message:
    properties:
      acknowledgedAt:
        type: string
      attempts:
        type: integer
      claimToken:
        type: string
      createdAt:
        type: string
      destinationStoreId:
        type: string
      id:
        type: string
      idEmpresa:
        type: string
      lastAttemptAt:
        type: string
      lastError:
        type: string
      leaseUntil:
        type: string
      nextAttemptAt:
        type: string
      operation:
        type: string
      originStoreId:
        type: string
      payload:
        type: object
      status:
        type: string
      syncBatchId:
        type: string
      syncJobId:
        type: string
      transferId:
        type: string
      type:
        type: string
      updatedAt:
        type: string
    type: object
  models.MessageDTO:
    properties:
      createdAt:
        type: string
      destinationStoreId:
        type: string
      id:
        type: string
      operation:
        maxLength: 100
        type: string
      payload:
        type: object
      type:
        enum:
        - EVENT
        - COMMAND
        type: string
    required:
    - destinationStoreId
    - id
    - operation
    - payload
    - type
    type: object
...
  models.SyncBatch:
    properties:
      batchNumber:
        type: integer
      createdAt:
        type: string
      errorCount:
        type: integer
      id:
        type: string
      idEmpresa:
        type: string
      isLast:
        type: boolean
      itemCount:
        type: integer
      jobId:
        type: string
      messageId:
        type: string
      processedCount:
        type: integer
      status:
        type: string
      totalItems:
        type: integer
      updatedAt:
        type: string
    type: object
...
info:
  contact: {}
  description: StoreLink integrates stores. Administrators use JWT Bearer; each FNTS_Server
    installation uses its own fixed X-API-Key and database-validated X-Agent-ID.
  title: StoreLink API
  version: "1.0"
paths:
  /agent/heartbeat:
    post:
      consumes:
      - application/json
      parameters:
      - description: Agent ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: Estado do agente
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.HeartbeatDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: Recebe estado recente do agente local
      tags:
      - Protocolo agente
  /agent/messages:
    post:
      consumes:
      - application/json
      description: Operação isolada por empresa. Identidade e autorização validadas
        pelo servidor.
      parameters:
      - description: Agent ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: Dados da operação
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.MessageDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: Publica evento ou comando idempotente
      tags:
      - Protocolo agente
  /agent/messages/{id}/ack:
    post:
      consumes:
      - application/json
      description: Operação isolada por empresa. Identidade e autorização validadas
        pelo servidor.
      parameters:
      - description: Agent ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: UUID do recurso
        in: path
        name: id
        required: true
        type: string
      - description: Dados da operação
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AckDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: Confirma somente após commit local
      tags:
      - Protocolo agente
  /agent/messages/{id}/fail:
    post:
      consumes:
      - application/json
      description: Operação isolada por empresa. Identidade e autorização validadas
        pelo servidor.
      parameters:
      - description: Agent ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: UUID do recurso
        in: path
        name: id
        required: true
        type: string
      - description: Dados da operação
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.FailDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: Registra falha com backoff limitado
      tags:
      - Protocolo agente
  /agent/messages/pending:
    get:
      description: Operação isolada por empresa. Identidade e autorização validadas
        pelo servidor.
      parameters:
      - description: Agent ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: Quantidade de 1 a 100
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.Message'
              type: array
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: Reserva mensagens com claim e lease
      tags:
      - Protocolo agente
  /agent/sync/jobs/{id}/batches:
    post:
      consumes:
      - application/json
      description: Operação isolada por empresa. Identidade e autorização validadas
        pelo servidor.
      parameters:
      - description: Agent ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: UUID do recurso
        in: path
        name: id
        required: true
        type: string
      - description: Dados da operação
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.BatchDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SyncBatch'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SyncBatch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: Recebe lote idempotente da carga
      tags:
      - Protocolo agente
...
  /integracao/v1/mensagens/{id}/confirmar:
    post:
      consumes:
      - application/json
      description: O FNTS_Server confirma somente depois de aplicar a mensagem na
        base local.
      parameters:
      - description: Store credential ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: UUID da mensagem
        in: path
        name: id
        required: true
        type: string
      - description: Resultado do processamento
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AckDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: Confirma uma mensagem do contrato de integraÃ§Ã£o
      tags:
      - Direct integration
  /integracao/v1/mensagens/{id}/falhar:
    post:
      consumes:
      - application/json
      description: Recoloca a mensagem em retry com backoff ou em dead-letter conforme
        a polÃ­tica do StoreLink.
      parameters:
      - description: Store credential ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: UUID da mensagem
        in: path
        name: id
        required: true
        type: string
      - description: Falha do processamento
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.FailDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: Registra falha de uma mensagem do contrato de integraÃ§Ã£o
      tags:
      - Direct integration
  /integracao/v1/mensagens/pendentes:
    get:
      description: O FNTS_Server reserva mensagens da fila local do StoreLink com
        claim e lease.
      parameters:
      - description: Store credential ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: Quantidade de 1 a 100
        in: query
        name: limit
        type: integer
      - description: Prefixo da operaÃ§Ã£o
        in: query
        name: operation
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.Message'
              type: array
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: Reserva mensagens do contrato de integraÃ§Ã£o
      tags:
      - Direct integration
...
securityDefinitions:
  AgentAPIKey:
    description: Per-agent fixed key shown once at provisioning or rotation.
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Bearer token obtained from /auth/login.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"